fix(dashboard): pormotion decimal value definition

+++ packages/admin/dashboard/src/routes/promotions/promotion-edit-details/components/edit-promotion-form/edit-promotion-details-form.tsx
@@ -299,7 +299,7 @@ export const EditPromotionDetailsForm = ({
                                 onChange(
                                   e.target.value === ""
                                     ? null
-                                    : parseInt(e.target.value)
+                                    : parseFloat(e.target.value)
                                 )
                               }}
                             />
